+++ test/test_packages/missing_export/package_a/lib/package_a.dart
@@ -3,3 +3,4 @@ library package_a;
 export 'types/class_a.dart';
 export 'types/class_c.dart' show ClassC;
 export 'types/private_types.dart';
+export 'types/hidden_build_method.dart' show PublicClass;