+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - fix: don't print color control sequences if no terminal is attached (e.g. when piping the output to a file)
 - fix: carry over entry points of type aliases to the aliased type (especially important for private types)
 - fix: naming of unary operators was wrong (unaryunary...)
+- fix: handling of top level functions in the context of show / hide in exports
 
 ## Version 0.22.0
 - fix: Fixes an issue if we have to deal with two types with the same name